# backend/openapi.yaml
openapi: 3.0.0
info:
  title: "Sentry CRM - Omnichannel Backend API"
  description: |
    Documentación oficial de la API para el backend de Sentry CRM.
    Esta API proporciona servicios para la gestión de compañías multi-tenant, usuarios, conversaciones, y más.
    Utiliza autenticación basada en JWT (Bearer Token) para las rutas protegidas.
  version: "1.0.0"

servers:
  - url: https://crm-sentry.onrender.com
    description: Production Server
  - url: http://localhost:4000
    description: Servidor de Desarrollo Local

# Define los esquemas de seguridad
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: cuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [USER, AGENT, ADMIN, MASTER]
        companyId:
          type: string
          format: cuid
          nullable: true
    Company:
      type: object
      properties:
        id:
          type: string
          format: cuid
        name:
          type: string
        status:
          type: string
          enum: [ACTIVE, TRIAL, OVERDUE, CANCELED, BANNED]
        isActive:
          type: boolean
    Conversation:
      type: object
      properties:
        id:
          type: string
          format: cuid
        subject:
          type: string
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
    Error:
      type: object
      properties:
        status:
          type: string
          example: "fail"
        message:
          type: string
          example: "Mensaje de error descriptivo."

# Define la seguridad global para las rutas que la requieran
security:
  - bearerAuth: []

# Definición de los Endpoints
paths:
  /api/onboarding:
    post:
      summary: Registra una nueva compañía y su administrador
      description: Endpoint público para el proceso de onboarding de un nuevo tenant.
      tags: [Onboarding]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                companyName:
                  type: string
                  example: "Mi Nueva Empresa"
                adminEmail:
                  type: string
                  format: email
                  example: "admin@nuevaempresa.com"
                adminPassword:
                  type: string
                  format: password
                  example: "contraseña-segura-123"
      responses:
        "201":
          description: Compañía y administrador creados exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                  companyId:
                    type: string
                  adminId:
                    type: string
        "400":
          description: Datos de entrada inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/auth/login:
    post:
      summary: Inicia sesión de un usuario
      description: Autentica a un usuario con email y contraseña y devuelve un token JWT.
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "master@sentrycrm.cloud"
                password:
                  type: string
                  format: password
                  example: "password123"
      responses:
        "200":
          description: Autenticación exitosa.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "401":
          description: Credenciales incorrectas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/users:
    get:
      summary: Obtiene una lista de todos los usuarios
      description: Ruta protegida que devuelve todos los usuarios del sistema.
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de usuarios obtenida con éxito.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  results:
                    type: integer
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: "#/components/schemas/User"
        "401":
          description: No autenticado.

  /api/users/{id}:
    get:
      summary: Obtiene un usuario por su ID
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Usuario encontrado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        $ref: "#/components/schemas/User"
        "404":
          description: Usuario no encontrado.
    patch:
      summary: Actualiza un usuario
      description: Permite a un usuario actualizar su propio perfil.
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Usuario actualizado.
        "403":
          description: No autorizado para editar este perfil.

  /api/admin/companies:
    get:
      summary: (MASTER) Obtiene todas las compañías
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de compañías.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
        "403":
          description: Acceso denegado.

  # ... (Y así sucesivamente para el resto de los endpoints de admin, posts, replies, conversations, etc.)

  /webhook:
    get:
      summary: Verificación del Webhook de Meta
      description: Endpoint para que Meta verifique la URL del webhook (handshake).
      tags: [Webhooks]
      parameters:
        - name: hub.mode
          in: query
          schema:
            type: string
        - name: hub.verify_token
          in: query
          schema:
            type: string
        - name: hub.challenge
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Verificación exitosa.
    post:
      summary: Recibe notificaciones del Webhook de Meta
      description: Endpoint que recibe los eventos de mensajería de la API de WhatsApp Cloud.
      tags: [Webhooks]
      responses:
        "200":
          description: Evento recibido.